# backend/src/config/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Signet Workforce ERP API",
    "version": "1.0.0",
    "description": "Production REST API for Signet Workforce ERP. Authenticate via POST /api/auth/login, then use the Bearer token for protected routes."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MonthYearBody": {
        "type": "object",
        "required": [
          "month",
          "year"
        ],
        "properties": {
          "month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12,
            "example": 6
          },
          "year": {
            "type": "integer",
            "minimum": 2000,
            "maximum": 2100,
            "example": 2026
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/attendance": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "List attendance records",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "employeeId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "siteId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "fromDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "toDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated attendance records"
          }
        }
      }
    },
...
    "/api/payroll/process": {
      "post": {
        "tags": [
          "Payroll"
        ],
        "summary": "Process payroll for a month",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/MonthYearBody"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "employeeIds": {
                        "type": "array",
                        "items": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payroll run id"
          }
        }
      }
    },
    "/api/payroll/payslips": {
      "get": {
        "tags": [
          "Salary Slips"
        ],
        "summary": "List salary slips",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "employeeId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated salary slips"
          }
        }
      }
    },
...
    "/api/reports/attendance": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Attendance summary report",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "fromDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "toDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance report"
          }
        }
      }
    },
    "/api/reports/payroll": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Payroll summary report",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payroll report"
          }
        }
      }
    },
    "/api/reports/billing": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Billing summary report",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Billing report"
          }
        }
      }
    }
  },
...
}